movido todas as requisições para APISERVICE
