+++ suduRecipes/suduRecipes/CellRecipe.xib
@@ -68,7 +68,7 @@
                 <outlet property="labelTextDescription" destination="yKj-4D-oQO" id="UdL-J2-WtJ"/>
                 <outlet property="labelTitle" destination="Dmt-hG-Y3p" id="MIx-E4-ddu"/>
             </connections>
-            <point key="canvasLocation" x="137.68115942028987" y="36.160714285714285"/>
+            <point key="canvasLocation" x="97" y="11"/>
         </tableViewCell>
     </objects>
 </document>
