+++ suduRecipes/suduRecipes/LoginViewController.swift
@@ -22,45 +22,22 @@ class LoginViewController: UIViewController {
         let email = textFieldEmail.text ?? ""
         let password = textFieldPassword.text ?? ""
         
-        //Troquei esse
-        //enterLogin(email: email, password: password)
-        
-        //Por esse
+        if email != "" && password != "" {
         APIService.enterLogin(email: email, password: password, completion: { responseData, error in
-            if error == nil, let response = responseData {
-                print(response)
-            } else {
+           
+            if error != nil {
+                print("error no acesso: \(String(describing: error))")
                 self.alertMessage(title: "Dados incorretos", message: "Digite novamente")
+                
             }
         })
+        } else {
+            self.alertMessage(title: "Dados incorretos", message: "email e Senha são requeridos!")
+        }
     }
-    
     override func viewDidLoad() {
         super.viewDidLoad()        
     }
-    
-    func enterLogin(email:String, password:String) {
-        let url = URL (string: "http://localhost:3003/users/login")!
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        
-        let json = [
-            "email": email,
-            "password": password
-        ]
-        
-        let jsonData = try! JSONSerialization.data(withJSONObject: json, options: [])
-       
-        let task = URLSession.shared.uploadTask(with: request, from: jsonData) { data, response, error in
-            print("response: \(String(describing: response))")
-            print("error: \(String(describing: error))")
-
-            guard let _ = data else {return}
-        }
-        task.resume()
-    }
-    
     override func viewWillAppear(_ animated: Bool) {
         self.navigationController?.setNavigationBarHidden(false, animated: false)
     }
