+++ suduRecipes/suduRecipes/SignupViewController.swift
@@ -36,13 +36,18 @@ class signupViewController: UIViewController {
         let password = textFieldPassword.text ?? ""
         let passwordConfirm = textFieldPasswordConfirm.text ?? ""
         
-        
-                        /*Validar senhas iguais*/
                         if password == passwordConfirm {
                             
                             if name != "" &&  email != "" && password != "" {
                                 
-                                createUsers(name: name,email: email,password: password)
+                                APIService.createUsers(name: name, email: email, password: password, completion:{ responseData, error in
+                                    if error == nil {
+                                        self.alertConfirm(title: "Sucesso", message: "Usuário cadastrado!")
+                                        
+                                    }else{
+                                        self.alertMessage(title: "Dados incorretos", message: "Digite novamente")
+                                    }
+                                })
                                
                             }else {
                                 self.alertMessage(title: "Dados incorretos", message: "Não pode deixar vazio, preencha todos os campos!")
@@ -50,53 +55,8 @@ class signupViewController: UIViewController {
                             
                         }else {
                             self.alertMessage(title: "Dados incorretos", message: "As senhas precisam ser iguais")
-                } /*fim valid senha*/
+                } 
             }
-        
-    
-    func createUsers(name:String, email:String, password:String) {
-        let url = URL (string: "http://localhost:3003/users/signup")!
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        
-        let json = [
-            "name": name,
-            "email": email,
-            "password": password
-        ]
-        
-        let jsonData = try! JSONSerialization.data(withJSONObject: json, options: [])
-       
-        let task = URLSession.shared.uploadTask(with: request, from: jsonData) { data, response, error in
-            print("response: \(String(describing: response))")
-            print("error: \(String(describing: error))")
-
-            guard let responseData = data else {return}
-        
-            do {
-                let jsonObject =  try JSONSerialization.jsonObject(with: responseData, options: []) as? [ String: Any]
-                if let dicionario = jsonObject {
-                    if let token = dicionario["token"] as? String{
-                        UserDefaults.standard.set(token, forKey: "token")
-                        self.alertConfirm(title: "Sucesso!", message: "Usuário cadastrado, Bem vind@!")
-                    }
-                  
-                }
-                
-            }catch{
-            
-            }
-            if let responseString = String(data: responseData,
-                encoding: .utf8){
-                print(responseString)
-            }
-        }
-
-        
-        task.resume()
-    }
-    
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -106,16 +66,6 @@ class signupViewController: UIViewController {
     override func viewWillAppear(_ animated: Bool) {
         self.navigationController?.setNavigationBarHidden(false, animated: false)
     }
-    
 
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
